requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.8
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
pillow==10.3.0
pycparser==2.22
PyMySQL==1.1.1
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
